[test-fopenfail-timeout-fix]: Set fopen-fail timeout to 300
Thanks to Víctor M. Jáquez L. <vjaquez@igalia.com> for the solution.

The problem is, if I understand it correctly, the timeout value:

By default the test timeout (tests/test_driver.pl) is 5 seconds, but in this
particular test, the timeout is reached before hitting the "too many open
files" error, which is the expected error. This is the case when the computer
is not under a heavy load.

This commit changes the timeout value only for the fail-open test to
300, allowing the tests to pass.

Signed-off-by: Manoj Srivastava <srivasta@golden-gryphon.com>