[master]: Make the binary packages M-A: Allowed
From: Helmut Grohne
it is doubtful wheter we can actually go
m-a:foreign. soon someone will come along and point out that the
"load" directive is not architecture agnosting and he'd be right.

the advantages are:
1) it's trivial to implement
2) after (cross)build-essential depends on make:any a large swath of
   things will just work
3) it's something that works today.

the single downside is that if we ever determine that m-a:foreign is
better, the move will be a bit painful.

Signed-off-by: Manoj Srivastava <srivasta@debian.org>