[master]: Increase fopen fail test timeout
Make has a test where we Recurse infinitely until we run out of open
files, and ensure we fail with a non-zero exit code., On some machines
the default time out was inadequate, since the test did not run till
conclusion. Doubling it did not have any effect. Quadrupling the
timeout worked.

Signed-off-by: Manoj Srivastava <srivasta@debian.org>